Move experimental currency formatting to legacy
Moving everything except mod.rs from currency to currency/legacy. Moved name_and_pattern from essentials.rs to legacy/mod.rs to avoid essentials depending on legacy code. Added TODO to each legacy file.
