+++ components/experimental/src/dimension/currency/legacy/long_compact_formatter.rs
@@ -2,6 +2,8 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
+// TODO: Legacy code. Should be removed after implementing the new implementation of https://hackmd.io/@younies/number_formatter_4x
+
 use core::fmt::Display;
 
 use fixed_decimal::Decimal;
@@ -10,12 +12,12 @@ use icu_plurals::PluralRules;
 use icu_provider::prelude::*;
 use writeable::Writeable;
 
-use crate::dimension::currency::compact_formatter::CompactCurrencyFormatterPreferences;
+use crate::dimension::currency::legacy::compact_formatter::CompactCurrencyFormatterPreferences;
 use crate::dimension::provider::currency::{
     extended::CurrencyExtendedDataV1, patterns::CurrencyPatternsDataV1,
 };
 
-use super::CurrencyCode;
+use crate::dimension::currency::CurrencyCode;
 
 extern crate alloc;
 
@@ -180,7 +182,7 @@ impl LongCompactCurrencyFormatter {
     ///
     /// # Examples
     /// ```
-    /// use icu::experimental::dimension::currency::long_compact_formatter::LongCompactCurrencyFormatter;
+    /// use icu::experimental::dimension::currency::legacy::long_compact_formatter::LongCompactCurrencyFormatter;
     /// use icu::experimental::dimension::currency::CurrencyCode;
     /// use icu::locale::locale;
     /// use tinystr::*;
