+++ components/experimental/src/dimension/currency/legacy/long_formatter.rs
@@ -2,6 +2,8 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
+// TODO: Legacy code. Should be removed after implementing the new implementation of https://hackmd.io/@younies/number_formatter_4x
+
 use core::fmt::Display;
 
 use fixed_decimal::Decimal;
@@ -14,7 +16,8 @@ use crate::dimension::provider::currency::{
     extended::CurrencyExtendedDataV1, patterns::CurrencyPatternsDataV1,
 };
 
-use super::{CurrencyCode, formatter::CurrencyFormatterPreferences};
+use crate::dimension::currency::CurrencyCode;
+use super::formatter::CurrencyFormatterPreferences;
 
 extern crate alloc;
 
@@ -147,7 +150,7 @@ impl LongCurrencyFormatter {
     ///
     /// # Examples
     /// ```
-    /// use icu::experimental::dimension::currency::long_formatter::LongCurrencyFormatter;
+    /// use icu::experimental::dimension::currency::legacy::long_formatter::LongCurrencyFormatter;
     /// use icu::experimental::dimension::currency::CurrencyCode;
     /// use icu::locale::locale;
     /// use tinystr::*;
